SVAT.py
# -*- coding: utf-8 -*-
from IPython.display import display
import ipywidgets as widgets
import snowflake.connector
import pandas as pd
from datetime import datetime
import time

# ========== SNOWFLAKE FUNCTIONS ==========
# [Keep all your existing Snowflake functions here unchanged]

# ========== NOTEBOOK UI ==========
class SnowflakeValidatorUI:
    def __init__(self):
        self.conn = None
        self.login_success = False
        
        # Create widgets
        self.create_login_ui()
        self.create_clone_ui()
        self.create_validation_ui()
        self.create_kpi_ui()
        
        # Arrange tabs
        self.tab = widgets.Tab()
        self.tab.children = [
            self.login_tab,
            self.clone_tab,
            self.validation_tab,
            self.kpi_tab
        ]
        self.tab.titles = ["🔐 Login", "⎘ Clone", "🔍 Validation", "📊 KPI"]
        display(self.tab)
        
        # Initially disable all but login tab
        self.set_tab_visibility()
    
    def create_login_ui(self):
        # Login widgets
        self.user = widgets.Text(placeholder="your_username", description="Username:")
        self.password = widgets.Password(placeholder="••••••••", description="Password:")
        self.account = widgets.Text(placeholder="account.region", description="Account:")
        self.login_btn = widgets.Button(description="Connect", button_style="success")
        self.disconnect_btn = widgets.Button(description="Disconnect", button_style="danger")
        self.status = widgets.Output()
        
        # Login layout
        self.login_tab = widgets.VBox([
            widgets.HTML("<h2>Snowflake Connection</h2>"),
            self.user,
            self.password,
            self.account,
            widgets.HBox([self.login_btn, self.disconnect_btn]),
            self.status
        ])
        
        # Event handlers
        self.login_btn.on_click(self.handle_login)
        self.disconnect_btn.on_click(self.handle_logout)
    
    def create_clone_ui(self):
        # Clone widgets
        self.source_db = widgets.Dropdown(description="Source DB:")
        self.source_schema = widgets.Dropdown(description="Source Schema:")
        self.target_schema = widgets.Text(description="Target Schema:")
        self.clone_btn = widgets.Button(description="Execute Clone", button_style="primary")
        self.clone_output = widgets.Output()
        
        # Clone layout
        self.clone_tab = widgets.VBox([
            widgets.HTML("<h2>Schema Clone</h2>"),
            widgets.HBox([
                widgets.VBox([
                    widgets.HTML("<h3>Source Selection</h3>"),
                    self.source_db,
                    self.source_schema,
                    self.target_schema,
                    self.clone_btn
                ]),
                widgets.VBox([
                    widgets.HTML("<h3>Clone Status</h3>"),
                    self.clone_output
                ])
            ])
        ])
        
        # Event handlers
        self.clone_btn.on_click(self.execute_clone)
        self.source_db.observe(self.update_schemas, names='value')
    
    # [Similar create_validation_ui() and create_kpi_ui() methods would go here]
    
    def handle_login(self, b):
        with self.status:
            self.status.clear_output()
            conn, msg = get_snowflake_connection(
                self.user.value, 
                self.password.value, 
                self.account.value
            )
            if conn:
                self.conn = conn
                self.login_success = True
                print(msg)
                self.update_database_dropdowns()
                self.set_tab_visibility()
            else:
                print(msg)
    
    # [Other methods would follow...]

# Create and display the UI
ui = SnowflakeValidatorUI()