requirements.txt.txt
streamlit==1.35.0
pandas==2.2.2
snowflake-connector-python==3.6.0
urllib3==1.26.18
